🔧 fix: GET 구조 설정
