+++ src/components/recruting/home/_main/MainContainer.tsx
@@ -1,8 +1,15 @@
 import MainCategory from "./MainCategory";
 import ClubCardList from "./ClubCardList";
 import Paging from "./Paging";
+import { useQuery } from "@tanstack/react-query";
+import { getMainClub } from "./service/mainClub";
 
 export default function MainContainer() {
+  const { data: mainClubsData } = useQuery<MainClubData>(
+    ["mainClubs"],
+    getMainClub
+  );
+
   return (
     <div>
       <div className="container max-w-[1077px] mx-auto">
@@ -10,9 +17,7 @@ export default function MainContainer() {
         <ClubCardList />
       </div>
 
-      <div className="pt-[96px] pb-[155px]">
-        <Paging />
-      </div>
+      <Paging />
     </div>
   );
 }
