+++ src/type/formtype.d.ts
@@ -51,19 +51,6 @@ declare interface PopularClubProps {
   tags: string[];
 }
 
-//메인 홈 동아리 리스트
-declare interface ClubCardProps {
-  id: string;
-  dDay: number;
-  clubImg: string;
-  logoSrc: string;
-  logoAlt: string;
-  title: string;
-  clubName: string;
-  tags: string[];
-  isLarge?: boolean; //크기 변환 prop
-}
-
 // 운영자 Type
 declare interface AdminUser {
   id: string;
